+++ src/embed_tests/TestPythonEngineProperties.cs
@@ -1,0 +1,70 @@
+using System;
+using NUnit.Framework;
+using Python.Runtime;
+
+namespace Python.EmbeddingTest
+{
+    public class TestPythonEngineProperties
+    {
+        [Test]
+        public static void GetBuildinfoDoesntCrash()
+        {
+            using (Py.GIL())
+            {
+                string s = PythonEngine.BuildInfo;
+
+                Assert.IsTrue(s.Length > 5);
+                Assert.IsTrue(s.Contains(","));
+            }
+        }
+
+        [Test]
+        public static void GetCompilerDoesntCrash()
+        {
+            using (Py.GIL())
+            {
+                string s = PythonEngine.Compiler;
+
+                Assert.IsTrue(s.Length > 0);
+                Assert.IsTrue(s.Contains("["));
+                Assert.IsTrue(s.Contains("]"));
+            }
+        }
+
+        [Test]
+        public static void GetCopyrightDoesntCrash()
+        {
+            using (Py.GIL())
+            {
+                string s = PythonEngine.Copyright;
+
+                Assert.IsTrue(s.Length > 0);
+                Assert.IsTrue(s.Contains("Python Software Foundation"));
+            }
+        }
+
+        [Test]
+        public static void GetPlatformDoesntCrash()
+        {
+            using (Py.GIL())
+            {
+                string s = PythonEngine.Platform;
+
+                Assert.IsTrue(s.Length > 0);
+                Assert.IsTrue(s.Contains("x") || s.Contains("win"));
+            }
+        }
+
+        [Test]
+        public static void GetVersionDoesntCrash()
+        {
+            using (Py.GIL())
+            {
+                string s = PythonEngine.Version;
+
+                Assert.IsTrue(s.Length > 0);
+                Assert.IsTrue(s.Contains(","));
+            }
+        }
+    }
+}