+++ CHANGELOG.md
@@ -51,6 +51,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Fixed `AppDomain` unload during GC (#397)(#400)
 -   Fixed `Py_Main` & `PySys_SetArgvEx` `no mem error` on `UCS4/PY3` (#399)
 -   Fixed `Python.Runtime.dll.config` on macOS (#120)
+-   Fixed crash on `PythonEngine.Version` (#413)
 
 ### Removed